인증 처리 완료

+++ src/main/java/site/metacoding/white/web/BoardApiController.java
@@ -36,6 +36,10 @@ public class BoardAPIController {
     @PostMapping("/board")
     public ResponseDto<?> save(@RequestBody BoardSaveReqDto boardSaveReqDto) {
         SessionUser sessionUser = (SessionUser) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            throw new RuntimeException("인증 필요");
+        }
+        // 핵심
         boardSaveReqDto.setSessionUser(sessionUser);
         BoardSaveRespDto boardSaveRespDto = boardService.save(boardSaveReqDto); // 서비스에는 단 하나의 객체만 전달한다.
         return new ResponseDto<>(1, "성공", boardSaveRespDto);
@@ -44,22 +48,34 @@ public class BoardAPIController {
     // 게시물 상세보기 = 보드 + 유저 + 커멘트
     @GetMapping("/board/{id}")
     public ResponseDto<?> findById(@PathVariable Long id) {
+        // 필요에 따라 인증 체크
         return new ResponseDto<>(1, "성공", boardService.findById(id)); // Entity -> JSON 변경 (MessageConverter)
     }
 
     @GetMapping("/board")
     public List<Board> findAll() {
+        // 필요에 따라 인증 체크
         return boardService.findAll();
     }
 
     @PutMapping("/board/{id}")
     public ResponseDto<?> update(@PathVariable Long id, @RequestBody BoardUpdateReqDto boardUpdateReqDto) {
+        SessionUser sessionUser = (SessionUser) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            throw new RuntimeException("인증 필요");
+        }
+        // 핵심
         boardUpdateReqDto.setId(id);
         return new ResponseDto<>(1, "성공", boardService.update(boardUpdateReqDto));
     }
 
     @DeleteMapping("/board/{id}")
     public ResponseDto<?> deleteById(@PathVariable Long id) {
+        SessionUser sessionUser = (SessionUser) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            throw new RuntimeException("인증 필요");
+        }
+        // 핵심
         boardService.deleteById(id);
         return new ResponseDto<>(1, "성공", null); // Entity -> JSON 변경 (MessageConverter)
     }
