+++ src/main/java/site/metacoding/white/web/CommentAPIController.java
@@ -24,12 +24,19 @@ public class CommentAPIController {
     @PostMapping("/comment")
     public ResponseDto<?> saveComment(@RequestBody CommentSaveReqDto commentSaveReqDto) {
         SessionUser user = (SessionUser) session.getAttribute("sessionUser");
+        if (user == null) { // 인증 체크를 위한 공통 로직
+            throw new RuntimeException("로그인 필요");
+        }
         commentSaveReqDto.setSessionUser(user);
         return new ResponseDto<>(1, "성공", commentService.save(commentSaveReqDto));
     }
 
     @DeleteMapping("/comment/{id}")
     public ResponseDto<?> deleteById(@PathVariable Long id) {
+        SessionUser user = (SessionUser) session.getAttribute("sessionUser");
+        if (user == null) {
+            throw new RuntimeException("로그인 필요");
+        }
         commentService.deleteById(id);
         return new ResponseDto<>(null, null, null);
     }
